% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean earthquake data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A dataframe containing earthquake information.}
}
\value{
The dataframe with cleaned date and location columns.
}
\description{
Takes in earthquake data, converts the year/month/day columns into a unified
date column, ensures the latitude and longitude are numeric, and cleans
up the location names.
}
\examples{
\dontrun{cleanData <- read_tsv("signif.txt") \%>\% eq_clean_data()}

}
