% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an informative pop up label.}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The earthquake data}
}
\value{
An HTML formatted text string.

\dontrun{
  cleanData %>% dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) %>% dplyr::mutate(popup_text = eq_create_label(.)) %>% eq_map(annot_col = "popup_text")
}
}
\description{
Takes a data frame containing earthquake data, and creates an informative HTML
label to be used on a leaflet map, containing magnitude, death toll, and location
information.
}
