% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean earthquake location name}
\usage{
eq_location_clean(locationName)
}
\arguments{
\item{locationName}{A location name to be cleaned.}
}
\value{
A location name in title case, without the country name.
}
\description{
Cleans up a passed location string to remove the initial country
name and colon, and convert to title case.
}
\examples{
hello_world(locationName = "UNITED KINGDOM: CHANNEL ISLANDS; GAUL")

}
