% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_map}
\alias{eq_map}
\title{Plot earthquakes on a leaflet map.}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{The earthquake data}

\item{annot_col}{The name of the column to display in the pop up.}
}
\value{
A leaflet map.
}
\description{
Takes a data frame containing earthquake data, as well as a specified column,
and creates a leaflet map plotting earthquakes with size based on their intensity, and
a pop up label containing the information in the specified column.
}
\examples{
\dontrun{
  cleanData \%>\% dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% eq_map(annot_col = "DATE")
}

}
