% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create a visualization of earthquakes on a timeline.}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xmin = NULL, xmax = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes or aes_. Must include 'x' values as dates. Deaths can be supplied as 'fill', and intensity as 'size.'  'y' can be used for a grouping variable.}

\item{data}{The data to be displayed in this layer. Must be earthquake data as specified above.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders.}

\item{xmin}{The earliest date to display}

\item{xmax}{The latest date to display}
}
\description{
This function creates a visualization layer for plotting earthquakes on a timeline, with
their magnitude as the size of the point, and their death toll as the colour intensity.
Optionally, separate timelines can be created by grouping, e.g., by country.
}
\examples{
\dontrun{
  cleanData \%>\% ggplot(aes(x = DATE, size=EQ_PRIMARY, fill = DEATHS, label = LOCATION_NAME)) + geom_timeline(alpha = 0.25, xmin = ymd('2000-01-01'), xmax = ymd('2015-12-31')) + guides(size = guide_legend(title = "Richter scale value"), fill = guide_colourbar(title = "# Deaths")) + theme_quakes()
}


}
