% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{GeomTimeLine}
\alias{GeomTimeLine}
\title{Plot earthquake data on a timeline.}
\format{An object of class \code{GeomTimeLine} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
GeomTimeLine
}
\value{
Returns a grid grob containing the earthquake timeline as a layer
}
\description{
Function that implements a class to visualize earthquake data, as
described in function \code{\link{geom_timeline}}.
}
\examples{
\dontrun{  ggplot2::layer(geom = GeomTimeLine, mapping = mapping, data = data, stat = stat,
position = position, show.legend = show.legend, inherit.aes = inherit.aes,
 params = list(na.rm = na.rm, nmax = nmax, xmin = xmin, xmax = xmax, ...))}
}
\keyword{datasets}
