% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_quakes}
\alias{theme_quakes}
\title{Theme for an earthquake timeline}
\usage{
theme_quakes()
}
\value{
A theme for an earthquake timeline.
}
\description{
Arranges the legend and other settings in a pleasing way for an earthquake timeline
created with \code{\link{geom_timeline}}.
}
\examples{
\dontrun{
  cleanData \%>\% ggplot(aes(x = DATE, size=EQ_PRIMARY, fill = DEATHS, label = LOCATION_NAME)) +
  geom_timeline(alpha = 0.25, xmin = ymd('2000-01-01'), xmax = ymd('2015-12-31')) +
  guides(size = guide_legend(title = "Richter scale value"),
  fill = guide_colourbar(title = "# Deaths")) + theme_quakes()
}


}
